+++ src/components/Table/components/CellRenderer/components/index.ts
@@ -1,4 +1,5 @@
 export * from './BooleanCellRenderer';
 export * from './DefaultCellRenderer';
+export * from './ImageCellRenderer';
 export * from './LayoutCellRenderer';
 export * from './PreformattedCellRenderer';