+++ src/migration.ts
@@ -4,6 +4,7 @@ import semver from 'semver';
 
 import { getColumnEditorConfig } from '@/utils';
 
+import { PAGE_SIZES } from './constants';
 import {
   ColumnAlignment,
   ColumnAppearanceConfig,
@@ -15,6 +16,7 @@ import {
   ColumnHeaderFontSize,
   ColumnPinDirection,
   ColumnSortConfig,
+  ImageScale,
   NestedObjectConfig,
   PaginationMode,
   PanelOptions,
@@ -291,6 +293,13 @@ export const getMigratedOptions = async (panel: PanelModel<OutdatedPanelOptions>
           normalized.preformattedStyle = false;
         }
 
+        /**
+         * Normalize scale fot items
+         */
+        if (panel.pluginVersion && semver.lt(panel.pluginVersion, '1.9.0') && !normalized.hasOwnProperty('scale')) {
+          normalized.scale = ImageScale.AUTO;
+        }
+
         return normalized;
       });
 
@@ -323,6 +332,7 @@ export const getMigratedOptions = async (panel: PanelModel<OutdatedPanelOptions>
           datasource: normalizeDatasourceOptions(dataSources, normalizedGroup.update.datasource),
         };
       }
+
       /**
        * Normalize Pagination
        */
@@ -330,6 +340,22 @@ export const getMigratedOptions = async (panel: PanelModel<OutdatedPanelOptions>
         normalizedGroup.pagination = {
           enabled: false,
           mode: PaginationMode.CLIENT,
+          defaultPageSize: PAGE_SIZES[0],
+        };
+      }
+
+      /**
+       * Normalize Pagination defaultPageSize
+       */
+      if (
+        panel.pluginVersion &&
+        semver.lt(panel.pluginVersion, '1.9.0') &&
+        normalizedGroup.pagination &&
+        !normalizedGroup.pagination.defaultPageSize
+      ) {
+        normalizedGroup.pagination = {
+          ...normalizedGroup.pagination,
+          defaultPageSize: PAGE_SIZES[0],
         };
       }
 
@@ -399,5 +425,6 @@ export const getMigratedOptions = async (panel: PanelModel<OutdatedPanelOptions>
 
     options.nestedObjects = nestedObjectsUpdated;
   }
+
   return options as PanelOptions;
 };